support perf to execute all speed tests
This commit will make `mc support perf <alias>` to execute all
performance tests, networking, drive and object speed tests.

perf will only have --duration and --verbose flags.

The user can still test with 'drive', 'object' and 'net' but they will
be hidden for now along with their flags.
